split off other components
